requirements.txt
Django==1.10.5
django-cors-headers==2.0.2
django-extensions==1.7.6
django-geojson==2.10.0
django-leaflet==0.20.0
djangorestframework==3.5.4
djangorestframework-gis==0.11
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
pyproj==1.9.5.1
six==1.10.0
